+++ packages/evokit-link/__tests__/Link.size.test.jsx
@@ -1,5 +1,7 @@
 import React from 'react';
 import { shallow } from 'enzyme';
+import barista from 'seed-barista';
+import style from '../style.css';
 import { Link } from '../src';
 
 describe('<Link link-size="..." />', () => {
@@ -8,5 +10,21 @@ describe('<Link link-size="..." />', () => {
         expect(wrapper).toMatchSnapshot();
 
         expect(wrapper.html()).toEqual('<a class="ek-link ek-link_size_value"></a>');
+    });
+
+    it('css', () => {
+        const css = barista({ content: style });
+
+        expect(css.rule('.ek-link_size_none').exists()).toBeTruthy();
+        expect(css.rule('.ek-link_size_inherit').exists()).toBeTruthy();
+        expect(css.rule('.ek-link_size_small').exists()).toBeTruthy();
+        expect(css.rule('.ek-link_size_default').exists()).toBeTruthy();
+        expect(css.rule('.ek-link_size_big').exists()).toBeTruthy();
+        expect(css.rule('.ek-link_size_h1').exists()).toBeTruthy();
+        expect(css.rule('.ek-link_size_h2').exists()).toBeTruthy();
+        expect(css.rule('.ek-link_size_h3').exists()).toBeTruthy();
+        expect(css.rule('.ek-link_size_h4').exists()).toBeTruthy();
+        expect(css.rule('.ek-link_size_h5').exists()).toBeTruthy();
+        expect(css.rule('.ek-link_size_h6').exists()).toBeTruthy();
     });
 });